requirements.txt
streamlit
python-docx
python-pptx
pdfplumber
pymupdf
translators
pandas
openpyxl
